requirements.txt
av==10.0.0
ctranslate2
faster-whisper
Flask==3.0.0
huggingface-hub==0.17.3
onnxruntime==1.16.1
openai==0.27.0
python-dotenv==0.19.1
tiktoken==0.3.3
tokenizers==0.14.1
torch==2.0.1